requirements.txt
amqp==5.1.1
asgiref==3.7.2
azure-core==1.29.4
azure-storage-blob==12.18.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.4
Django==3.2
django-cors-headers==3.7.0
django-filter==23.3
django-pwa==1.1.0
django-storages==1.14.1
djangorestframework==3.12.4
ElasticEmail==4.0.23
gunicorn==20.1.0
idna==3.4
isodate==0.6.1
kombu==5.3.2
mssql-django==1.3
prompt-toolkit==3.0.39
pycparser==2.21
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==4.6.0
requests==2.31.0
sib-api-v3-sdk==7.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.8
